+++ lib/main.ts
@@ -1,11 +1,11 @@
 export { ReactGrid } from "./components/ReactGrid";
 
 export type * from "./types/PublicModel";
-export type * from "./types/CellMatrix";
 
 export { TextCell } from "./cellTemplates/TextCell";
 export { NumberCell } from "./cellTemplates/NumberCell";
-export { DateCell } from "./cellTemplates/DateCell";
-export { HeaderCell } from "./cellTemplates/HeaderCell";
+
+export { default as CellWrapper } from "./components/CellWrapper";
+export { useCellContext } from "./components/CellContext";
 
 export * from "./hooks/useReactGridAPI";
