+++ lib/types/ReactGridStore.ts
@@ -43,6 +43,8 @@ export interface ReactGridStoreProps {
   shadowPosition?: number;
   shadowSize?: number;
 
+  empty?: undefined;
+
   onFillHandle?: (selectedArea: NumericalRange, fillRange: NumericalRange) => void;
   onAreaSelected?: (selectedArea: NumericalRange) => void;
   onCellFocused?: (cellLocation: IndexedLocation) => void;
