+++ stories/FillHandle.stories.tsx
@@ -1,11 +1,12 @@
 import { StrictMode, useState } from "react";
 import { cellMatrixBuilder } from "../lib/utils/cellMatrixBuilder";
-import { Column, Row, HeaderCell, TextCell, ReactGrid, NumberCell } from "../lib/main";
+import { Column, Row, TextCell, ReactGrid, NumberCell } from "../lib/main";
 import { StoryDefault } from "@ladle/react";
 import { ErrorBoundary } from "../lib/components/ErrorBoundary";
 import React from "react";
 import { initialColumns, initialRows, CellData, headerRow, dataRows, rgStyles } from "./utils/examplesConfig";
 import { handleFill } from "./utils/handleFill";
+import { HeaderCell } from "./cellTemplates/HeaderCell";
 
 export const FillHandleExample = () => {
   const [columns, setColumns] = useState<Array<Column>>(initialColumns);
